feat: Log system
